+++ tests/enums.spec.js
@@ -4,7 +4,7 @@ const enums = require('../src/enums')
 
 module.exports = function () {
   test('enums', (t) => {
-    t.plan(12)
+    t.plan(13)
 
     try {
       t.equal(enums.priorityFromValue(60), 'lowest', 'Priority from value 60 returns lowest')
@@ -13,12 +13,13 @@ module.exports = function () {
       t.equal(enums.priorityFromValue(30), 'medium', 'Priority from value 30 returns medium')
       t.equal(enums.priorityFromValue(20), 'high', 'Priority from value 20 returns high')
       t.equal(enums.priorityFromValue(10), 'highest', 'Priority from value 10 returns highest')
-      t.equal(Object.keys(enums.priority).length, 6, 'Enums priority has has correct number of keys')
+      t.equal(Object.keys(enums.state).length, 3, 'Enums state has the correct number of keys')
+      t.equal(Object.keys(enums.priority).length, 6, 'Enums priority has correct number of keys')
       t.equal(Object.keys(enums.status).length, 25, 'Enums status has correct number of keys')
       t.equal(Object.keys(enums.options).length, 11, 'Enums options has correct number of keys')
-      t.equal(Object.keys(enums.index).length, 3, 'Enums index has has correct number of keys')
-      t.equal(Object.keys(enums.log).length, 3, 'Enums log has has correct number of keys')
-      t.equal(Object.keys(enums.message).length, 13, 'Enums message has has correct number of keys')
+      t.equal(Object.keys(enums.index).length, 3, 'Enums index has correct number of keys')
+      t.equal(Object.keys(enums.log).length, 3, 'Enums log has correct number of keys')
+      t.equal(Object.keys(enums.message).length, 14, 'Enums message has correct number of keys')
     } catch (err) {
       tError(err, module, t)
     }