+++ tests/queue-interruption.spec.js
@@ -1,6 +1,7 @@
 const test = require('tape')
 const Promise = require('bluebird')
 const enums = require('../src/enums')
+const is = require('../src/is')
 const tError = require('./test-error')
 const Queue = require('../src/queue')
 const tOpts = require('./test-options')
@@ -12,7 +13,7 @@ const queueInterruption = proxyquire('../src/queue-interruption',
 module.exports = function () {
   return new Promise((resolve, reject) => {
     test('queue-interruption', (t) => {
-      t.plan(10)
+      t.plan(19)
 
       const q = new Queue(tOpts.cxn(), tOpts.default())
       processStub.restart = function (q) {
@@ -21,19 +22,34 @@ module.exports = function () {
 
       // ---------- Event Handler Setup ----------
       let testEvents = false
-      function pausingEventHandler (qId) {
+      let pausingEventCount = 0
+      let pausingEventTotal = 1
+      function pausingEventHandler (global, queueId) {
+        pausingEventCount++
         if (testEvents) {
-          t.equal(qId, q.id, `Event: pausing [${qId}]`)
+          t.pass(`Event: pausing [${pausingEventCount} of ${pausingEventTotal}] `)
+          t.ok(is.boolean(global), `Event: pausing [global: ${global}]`)
+          t.ok(is.string(queueId), `Event: pausing [queueId: ${queueId}]`)
         }
       }
-      function pausedEventHandler (qId) {
+      let pausedEventCount = 0
+      let pausedEventTotal = 1
+      function pausedEventHandler (global, queueId) {
+        pausedEventCount++
         if (testEvents) {
-          t.equal(qId, q.id, `Event: paused [${qId}]`)
+          t.pass(`Event: paused [${pausedEventCount} of ${pausedEventTotal}] `)
+          t.ok(is.boolean(global), `Event: paused [global: ${global}]`)
+          t.ok(is.string(queueId), `Event: paused [queueId: ${queueId}]`)
         }
       }
-      function resumedEventHandler (qId) {
+      let resumedEventCount = 0
+      let resumedEventTotal = 1
+      function resumedEventHandler (global, queueId) {
+        resumedEventCount++
         if (testEvents) {
-          t.equal(qId, q.id, `Event: resumed [${qId}]`)
+          t.pass(`Event: resumed [${resumedEventCount} of ${resumedEventTotal}] `)
+          t.ok(is.boolean(global), `Event: resumed [global: ${global}]`)
+          t.ok(is.string(queueId), `Event: resumed [queueId: ${queueId}]`)
         }
       }
       function addEventHandlers () {
@@ -73,6 +89,13 @@ module.exports = function () {
         t.ok(resumed, 'Interruption resume returns true')
         t.notOk(q.paused, 'Queue is not paused')
         removeEventHandlers()
+
+        // ---------- Event summary Test ----------
+        t.comment('queue-interruption: Event Summary')
+        t.equal(pausingEventCount, pausingEventTotal, 'Pausing event count valid')
+        t.equal(pausedEventCount, pausedEventTotal, 'Paused event count valid')
+        t.equal(resumedEventCount, resumedEventTotal, 'Resumed event count valid')
+
         q.stop()
         return resolve(t.end())
       }).catch(err => tError(err, module, t))