+++ tests/queue-state.spec.js
@@ -12,7 +12,7 @@ const queueState = require('../src/queue-state')
 module.exports = function () {
   return new Promise((resolve, reject) => {
     test('queue-state', (t) => {
-      t.plan(9)
+      t.plan(15)
 
       const q = new Queue(tOpts.cxn(), tOpts.default())
       const q2 = new Queue(tOpts.cxn(), tOpts.default())
@@ -21,33 +21,45 @@ module.exports = function () {
 
       // ---------- Event Handler Setup ----------
       let testEvents = false
+      let pausingEventCount = 0
+      let pausingEventTotal = 1
+      function pausingEventHandler (global, queueId) {
+        pausingEventCount++
+        if (testEvents) {
+          t.pass(`Event: pausing [${pausingEventCount} of ${pausingEventTotal}] `)
+          t.ok(is.boolean(global), `Event: pausing [global: ${global}]`)
+          t.ok(is.string(queueId), `Event: pausing [queueId: ${queueId}]`)
+        }
+      }
       let pausedEventCount = 0
-      let pausedEventTotal = 10
+      let pausedEventTotal = 1
       function pausedEventHandler (global, queueId) {
         pausedEventCount++
         if (testEvents) {
-          t.ok(pausedEventCount < 10, `Event: paused [${pausedEventCount} of ${pausedEventTotal}] `)
+          t.pass(`Event: paused [${pausedEventCount} of ${pausedEventTotal}] `)
           t.ok(is.boolean(global), `Event: paused [global: ${global}]`)
           t.ok(is.string(queueId), `Event: paused [queueId: ${queueId}]`)
         }
       }
       let resumedEventCount = 0
-      let resumedEventTotal = 10
+      let resumedEventTotal = 1
       function resumedEventHandler (global, queueId) {
         resumedEventCount++
         if (testEvents) {
-          t.ok(resumedEventCount < 10, `Event: resumed [${resumedEventCount} of ${resumedEventTotal}] `)
+          t.pass(`Event: resumed [${resumedEventCount} of ${resumedEventTotal}] `)
           t.ok(is.boolean(global), `Event: resumed [global: ${global}]`)
           t.ok(is.string(queueId), `Event: resumed [queueId: ${queueId}]`)
         }
       }
       function addEventHandlers () {
         testEvents = true
+        q.on(enums.status.paused, pausingEventHandler)
         q.on(enums.status.paused, pausedEventHandler)
         q.on(enums.status.resumed, resumedEventHandler)
       }
       function removeEventHandlers () {
         testEvents = false
+        q.removeListener(enums.status.paused, pausingEventHandler)
         q.removeListener(enums.status.paused, pausedEventHandler)
         q.removeListener(enums.status.resumed, resumedEventHandler)
       }
@@ -67,6 +79,12 @@ module.exports = function () {
         return queueState(q2, enums.status.active)
       }).then((resetResult) => {
         t.notOk(q.paused, 'Local queue is resumed')
+
+        // ---------- Event summary Test ----------
+        t.comment('queue-state: Event Summary')
+        t.equal(pausingEventCount, pausingEventTotal, 'Pausing event count valid')
+        t.equal(pausedEventCount, pausedEventTotal, 'Paused event count valid')
+        t.equal(resumedEventCount, resumedEventTotal, 'Resumed event count valid')
 
         removeEventHandlers()
         q.stop()