+++ tests/queue.spec.js
@@ -9,7 +9,7 @@ const Queue = require('../src/queue')
 module.exports = function () {
   return new Promise((resolve, reject) => {
     test('queue', (t) => {
-      t.plan(110)
+      t.plan(154)
 
       let q = new Queue(tOpts.cxn(), tOpts.queueNameOnly())
       let q2
@@ -29,70 +29,196 @@ module.exports = function () {
 
       // ---------- Event Handler Setup ----------
       let testEvents = false
-      function readyEventHandler (qId) {
+      let readyEventCount = 0
+      let readyEventTotal = 0
+      function readyEventHandler (queueId) {
+        readyEventCount++
         if (testEvents) {
-          t.ok(is.string(qId), `Event: ready [${qId}]`)
+          t.ok(is.string(queueId), `Event: ready [${readyEventCount} of ${readyEventTotal}] [${queueId}]`)
         }
       }
+      let addedEventCount = 0
+      let addedEventTotal = 4
       function addedEventHandler (jobId) {
+        addedEventCount++
         if (testEvents) {
-          t.ok(is.uuid(jobId), `Event: added [${jobId}]`)
+          t.ok(is.uuid(jobId), `Event: added [${addedEventCount} of ${addedEventTotal}] [${jobId}]`)
         }
       }
+      let activeEventCount = 0
+      let activeEventTotal = 3
+      function activeEventHandler (jobId) {
+        activeEventCount++
+        if (testEvents) {
+          t.ok(is.uuid(jobId), `Event: active [${activeEventCount} of ${activeEventTotal}] [${jobId}]`)
+        }
+      }
+      let processingEventCount = 0
+      let processingEventTotal = 3
       function processingEventHandler (jobId) {
+        processingEventCount++
         if (testEvents) {
-          t.ok(is.uuid(jobId), `Event: processing [${jobId}]`)
+          t.ok(is.uuid(jobId), `Event: processing [${processingEventCount} of ${processingEventTotal}] [${jobId}]`)
         }
       }
-      function pausedEventHandler (qId) {
+      let progressEventCount = 0
+      let progressEventTotal = 0
+      function progressEventHandler (jobId, percent) {
+        progressEventCount++
         if (testEvents) {
-          t.ok(is.string(qId), `Event: paused [${qId}]`)
+          t.pass(`Event: progress  [${progressEventCount} of ${progressEventTotal}]`)
+          t.ok(is.uuid(jobId), `Event: progress [${jobId}]`)
+          t.ok(is.number(percent), `Event: progress [${percent}%]`)
         }
       }
-      function resumedEventHandler (qId) {
+      let completedEventCount = 0
+      let completedEventTotal = 3
+      function completedEventHandler (jobId) {
+        completedEventCount++
         if (testEvents) {
-          t.ok(is.string(qId), `Event: resumed [${qId}]`)
+          t.ok(is.uuid(jobId), `Event: completed [${completedEventCount} of ${completedEventTotal}] [${jobId}]`)
         }
       }
+      let cancelledEventCount = 0
+      let cancelledEventTotal = 1
+      function cancelledEventHandler (jobId) {
+        cancelledEventCount++
+        if (testEvents) {
+          t.ok(is.uuid(jobId), `Event: cancelled [${cancelledEventCount} of ${cancelledEventTotal}] [${jobId}]`)
+        }
+      }
+      let failedEventCount = 0
+      let failedEventTotal = 0
+      function failedEventHandler (jobId) {
+        failedEventCount++
+        if (testEvents) {
+          t.ok(is.uuid(jobId), `Event: failed [${failedEventCount} of ${failedEventTotal}] [${jobId}]`)
+        }
+      }
+      let terminatedEventCount = 0
+      let terminatedEventTotal = 0
+      function terminatedEventHandler (jobId) {
+        terminatedEventCount++
+        if (testEvents) {
+          t.ok(is.uuid(jobId), `Event: terminated [${terminatedEventCount} of ${terminatedEventTotal}] [${jobId}]`)
+        }
+      }
+      let logEventCount = 0
+      let logEventTotal = 0
+      function logEventHandler (jobId) {
+        logEventCount++
+        if (testEvents) {
+          t.ok(is.uuid(jobId), `Event: log [${logEventCount} of ${logEventTotal}] [${jobId}]`)
+        }
+      }
+      let updatedEventCount = 0
+      let updatedEventTotal = 0
+      function updatedEventHandler (jobId) {
+        updatedEventCount++
+        if (testEvents) {
+          t.ok(is.uuid(jobId), `Event: updated [${updatedEventCount} of ${updatedEventTotal}] [${jobId}]`)
+        }
+      }
+      let pausingEventCount = 0
+      let pausingEventTotal = 2
+      function pausingEventHandler (global, queueId) {
+        pausingEventCount++
+        if (testEvents) {
+          t.ok(is.string(queueId), `Event: pausing [${pausingEventCount} of ${pausingEventTotal}]`)
+          t.ok(is.boolean(global), `Event: pausing [global: ${global}]`)
+          t.ok(is.string(queueId), `Event: pausing [queueId: ${queueId}]`)
+        }
+      }
+      let pausedEventCount = 0
+      let pausedEventTotal = 2
+      function pausedEventHandler (global, queueId) {
+        pausedEventCount++
+        if (testEvents) {
+          t.ok(is.string(queueId), `Event: paused [${pausedEventCount} of ${pausedEventTotal}]`)
+          t.ok(is.boolean(global), `Event: paused [global: ${global}]`)
+          t.ok(is.string(queueId), `Event: paused [queueId: ${queueId}]`)
+        }
+      }
+      let resumedEventCount = 0
+      let resumedEventTotal = 1
+      function resumedEventHandler (global, queueId) {
+        resumedEventCount++
+        if (testEvents) {
+          t.ok(is.string(queueId), `Event: resumed [${resumedEventCount} of ${resumedEventTotal}]`)
+          t.ok(is.boolean(global), `Event: resumed [global: ${global}]`)
+          t.ok(is.string(queueId), `Event: resumed [queueId: ${queueId}]`)
+        }
+      }
+      let removedEventCount = 0
+      let removedEventTotal = 1
       function removedEventHandler (jobId) {
+        removedEventCount++
         if (testEvents) {
-          t.ok(is.uuid(jobId), `Event: removed [${jobId}]`)
+          t.ok(is.uuid(jobId), `Event: removed [${removedEventCount} of ${removedEventTotal}] [${jobId}]`)
         }
       }
-      function idleEventHandler (qId) {
+      let idleEventCount = 0
+      let idleEventTotal = 4
+      function idleEventHandler (queueId) {
+        idleEventCount++
         if (testEvents) {
-          t.ok(is.string(qId), `Event: idle [${qId}]`)
-          q.removeListener(enums.status.idle, idleEventHandler)
+          t.ok(is.string(queueId), `Event: idle [${idleEventCount} of ${idleEventTotal}] [${queueId}]`)
         }
       }
+      let reviewedEventCount = 0
+      let reviewedEventTotal = 0
+      function reviewedEventHandler (queueId) {
+        reviewedEventCount++
+        if (testEvents) {
+          t.ok(is.string(queueId), `Event: reviewed [${reviewedEventCount} of ${reviewedEventTotal}] [${queueId}]`)
+        }
+      }
+      let resetEventCount = 0
+      let resetEventTotal = 1
       function resetEventHandler (total) {
+        resetEventCount++
         if (testEvents) {
-          t.ok(is.integer(total), `Event: reset [${total}]`)
+          t.ok(is.integer(total), `Event: reset [${resetEventCount} of ${resetEventTotal}] [${total}]`)
         }
       }
+      let errorEventCount = 0
+      let errorEventTotal = 3
       function errorEventHandler (err) {
+        errorEventCount++
         if (testEvents) {
-          t.ok(is.string(err.message), `Event: error [${err.message}]`)
+          t.ok(is.string(err.message), `Event: error [${errorEventCount} of ${errorEventTotal}] [${err.message}]`)
         }
       }
-      function detachedEventHandler (qId) {
+      let detachedEventCount = 0
+      let detachedEventTotal = 1
+      function detachedEventHandler (queueId) {
+        detachedEventCount++
         if (testEvents) {
-          t.ok(is.string(qId), `Event: detached [${qId}]`)
+          t.ok(is.string(queueId), `Event: detached [${detachedEventCount} of ${detachedEventTotal}] [${queueId}]`)
         }
       }
-      function stoppingEventHandler (qId) {
+      let stoppingEventCount = 0
+      let stoppingEventTotal = 1
+      function stoppingEventHandler (queueId) {
+        stoppingEventCount++
         if (testEvents) {
-          t.ok(is.string(qId), `Event: stopping [${qId}]`)
+          t.ok(is.string(queueId), `Event: stopping [${stoppingEventCount} of ${stoppingEventTotal}] [${queueId}]`)
         }
       }
-      function stoppedEventHandler (qId) {
+      let stoppedEventCount = 0
+      let stoppedEventTotal = 1
+      function stoppedEventHandler (queueId) {
+        stoppedEventCount++
         if (testEvents) {
-          t.ok(is.string(qId), `Event: stopped [${qId}]`)
+          t.ok(is.string(queueId), `Event: stopped [${stoppedEventCount} of ${stoppedEventTotal}] [${queueId}]`)
         }
       }
-      function droppedEventHandler (qId) {
+      let droppedEventCount = 0
+      let droppedEventTotal = 1
+      function droppedEventHandler (queueId) {
+        droppedEventCount++
         if (testEvents) {
-          t.ok(is.string(qId), `Event: dropped[${qId}]`)
+          t.ok(is.string(queueId), `Event: dropped [${droppedEventCount} of ${droppedEventTotal}] [${queueId}]`)
         }
       }
 
@@ -100,11 +226,21 @@ module.exports = function () {
         testEvents = true
         q.on(enums.status.ready, readyEventHandler)
         q.on(enums.status.added, addedEventHandler)
+        q.on(enums.status.active, activeEventHandler)
         q.on(enums.status.processing, processingEventHandler)
+        q.on(enums.status.progress, progressEventHandler)
+        q.on(enums.status.completed, completedEventHandler)
+        q.on(enums.status.cancelled, cancelledEventHandler)
+        q.on(enums.status.failed, failedEventHandler)
+        q.on(enums.status.terminated, terminatedEventHandler)
         q.on(enums.status.paused, pausedEventHandler)
+        q.on(enums.status.pausing, pausingEventHandler)
         q.on(enums.status.resumed, resumedEventHandler)
         q.on(enums.status.removed, removedEventHandler)
+        q.on(enums.status.log, logEventHandler)
+        q.on(enums.status.updated, updatedEventHandler)
         q.on(enums.status.idle, idleEventHandler)
+        q.on(enums.status.reviewed, reviewedEventHandler)
         q.on(enums.status.reset, resetEventHandler)
         q.on(enums.status.error, errorEventHandler)
         q.on(enums.status.detached, detachedEventHandler)
@@ -116,11 +252,21 @@ module.exports = function () {
         testEvents = false
         q.removeListener(enums.status.ready, readyEventHandler)
         q.removeListener(enums.status.added, addedEventHandler)
+        q.removeListener(enums.status.active, activeEventHandler)
         q.removeListener(enums.status.processing, processingEventHandler)
+        q.removeListener(enums.status.progress, progressEventHandler)
+        q.removeListener(enums.status.completed, completedEventHandler)
+        q.removeListener(enums.status.cancelled, cancelledEventHandler)
+        q.removeListener(enums.status.failed, failedEventHandler)
+        q.removeListener(enums.status.terminated, terminatedEventHandler)
         q.removeListener(enums.status.paused, pausedEventHandler)
+        q.removeListener(enums.status.pausing, pausingEventHandler)
         q.removeListener(enums.status.resumed, resumedEventHandler)
         q.removeListener(enums.status.removed, removedEventHandler)
+        q.removeListener(enums.status.log, logEventHandler)
+        q.removeListener(enums.status.updated, updatedEventHandler)
         q.removeListener(enums.status.idle, idleEventHandler)
+        q.removeListener(enums.status.reviewed, reviewedEventHandler)
         q.removeListener(enums.status.reset, resetEventHandler)
         q.removeListener(enums.status.error, errorEventHandler)
         q.removeListener(enums.status.detached, detachedEventHandler)
@@ -355,6 +501,33 @@ module.exports = function () {
         t.equal(jobCheck[0].status, enums.status.completed, 'Job is completed')
 
         removeEventHandlers()
+
+        // ---------- Event summary Test ----------
+        t.comment('queue: Event Summary')
+        t.equal(readyEventCount, readyEventTotal, 'Ready event count valid')
+        t.equal(addedEventCount, addedEventTotal, 'Added event count valid')
+        t.equal(activeEventCount, activeEventTotal, 'Active event count valid')
+        t.equal(processingEventCount, processingEventTotal, 'Processing event count valid')
+        t.equal(progressEventCount, progressEventTotal, 'Progress event count valid')
+        t.equal(completedEventCount, completedEventTotal, 'Completed event count valid')
+        t.equal(cancelledEventCount, cancelledEventTotal, 'Cancelled event count valid')
+        t.equal(failedEventCount, failedEventTotal, 'Failed event count valid')
+        t.equal(terminatedEventCount, terminatedEventTotal, 'Terminated event count valid')
+        t.equal(pausingEventCount, pausingEventTotal, 'Pausing event count valid')
+        t.equal(pausedEventCount, pausedEventTotal, 'Paused event count valid')
+        t.equal(resumedEventCount, resumedEventTotal, 'Resumed event count valid')
+        t.equal(removedEventCount, removedEventTotal, 'Removed event count valid')
+        t.equal(logEventCount, logEventTotal, 'Log event count valid')
+        t.equal(updatedEventCount, updatedEventTotal, 'Updated event count valid')
+        t.equal(idleEventCount, idleEventTotal, 'Idle event count valid')
+        t.equal(reviewedEventCount, reviewedEventTotal, 'Reviewed event count valid')
+        t.equal(resetEventCount, resetEventTotal, 'Reset event count valid')
+        t.equal(errorEventCount, errorEventTotal, 'Error event count valid')
+        t.equal(detachedEventCount, detachedEventTotal, 'Detached event count valid')
+        t.equal(stoppingEventCount, stoppingEventTotal, 'Stopping event count valid')
+        t.equal(stoppedEventCount, stoppedEventTotal, 'Stopped event count valid')
+        t.equal(droppedEventCount, droppedEventTotal, 'Dropped event count valid')
+
         q.stop()
         q2.stop()
         return resolve(t.end())